Title: 비교에 따른 연산, Time: 106ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240818/비교에 따른 연산/operation-based-on-comparison.py
@@ -4,4 +4,4 @@ if a>b:
     print(a*b)
 
 else:
-    print(b/a)
+    print(int(b/a))
